save ids to .txt file, add page to show .txt file
This commit saves every submission in a text file in the local
directory. It also adds a button that links to a page that views the
text file and has a button to return to the home page and clear the text
file to allow for a fresh scanning session.